backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# LLM Orchestration
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.12.0

# Data Management
pandas>=2.2.0
psycopg2-binary>=2.9.0

# Scraping
apify-client>=1.6.0

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
